code refactoring 🪛
